# Conflicts:
#	domain/src/test/java/com/github/anrimian/musicplayer/domain/business/TestBusinessDataProvider.java
#	presentation/src/main/java/com/github/anrimian/musicplayer/ui/common/dialogs/input/InputTextDialogFragment.java

+++ domain/src/test/java/com/github/anrimian/musicplayer/domain/business/TestBusinessDataProvider.java
@@ -35,9 +35,9 @@ public class TestBusinessDataProvider {
 
     private static List<PlayListItem> getFakePlayListItems() {
         List<PlayListItem> items = new ArrayList<>();
-        for (int i = 0; i < 100000; i++) {
+        for (long i = 0; i < 100000; i++) {
             Composition composition = fakeComposition(i, "music-" + i);
-            PlayListItem item = new PlayListItem(i, (long) i, composition);
+            PlayListItem item = new PlayListItem(i, i, composition);
             items.add(item);
         }
         return items;
